Basic functionality to change card Points
Because the call to the API is the same to add or remove points, these have been rolled into a single function that can take in either a positive or negative integer to add or subtract points, respectively. This could be changed in the future if it seems like separate functions would be more intuitive, eg 'addPoints' and 'removePoints'.