Return Promise from 'createContact' function
Instead of completely wrapping the request-promise call inside the 'createContact' function, use 'createContact' to return a customized rp call, ie an ES5-friendly Promise.

Updates to in-code documentation and to sample function call provided in commented-out code at the bottom to reflect that the function returns a promise. Added note re need to handle 'undefined' for optional properties.