Remove .env access from library (client responsibility)
The library should not access environment variables directly. That could easily be inconvenient for the client -- for example they may have some other environment variable called 'ACCESS_TOKEN'. Instead, let it be the client's responsibility to pass in their access token from wherever they like when initializing a new Lightrail().

This commit updates the Lightrail constructor so that it expects to receive the access token as an argument. It then establishes the BASE_URL and HEADERS as local variables (which should also be treated as constants). Associated methods have been updated to refer to the new location of the base url and the headers.

The 'dotenv' dependency has been removed from the package.json (no longer required for the library) but added into the testApp file, where it is now used from the 'client side'. The readme has bee updated to reflect these changes."