Placeholder test
This commit creates a placeholder test that ensures the lightrailClient spec can read the right files and variables ('Lightrail' constructor, env variable for access token).

A note about the access token: tests ultimately should not rely on the real Lightrail API, but for initial iterations they will. Until that is changed, they will need access to a real access token.